Manage multiple persistent profile infos
JProfiling results in the addition of profiling
information for a greater number of jitted bodies.
However, it was not possible to easy differientiate
between profiling information generated by prior
compilations or generated for profiling instrumentation
in the current. Moreover, persistent profile information
was not removed when it was no longer needed.

This change shifts profile information from the method
info to the body info, so that it can be cleaned up once
the body will no longer execute. Moreover, it adds tracking
on the method info to identify the most recent profiling
information and the best, with logic to replace the best
once the recent superceeds it.

To ensure profile information is cleaned up appropriately,
a reference count has been added to the structure, capable
of identifying when it is no longer needed and removing it.

Signed-off-by: Nicholas Coughlin <cnic@ca.ibm.com>