Scan JNI Weak Global References as a hard root for CS
JNI Weak Global References table is usually treated as Clearable root.
This means it is scanned at the end of Scavenger cycle. However this a
problem for Concurrent Scavenger because VM code does not call Read
Barrier on dereferencing elements of this table. As a result references
might be not fixed up properly at time of usage and this is a reason for
crashes in JNI support related code (JVMTI implementation for instance).
So each single place in the code where elements of the table are
dereferenced should be wrapped with macro calling Read Barrier. And
there are many of them.
The temporary solution (until all places required modification are
spotted) is treat JNI Weak Global References table as a hard root at the
time when Concurrent Scavenger cycle active. Concurrent Scavenger scans
all hard roots at the time of first increment so references in the table
would be fixed up before mutator threads have access to it.

Signed-off-by: dmitripivkine <Dmitri_Pivkine@ca.ibm.com>