Release VM access when waiting for messages from the server
Holding VMaccess prevents the GC from cleaning the heap and may also
prevent other activities (like class redefinition) from happening.
The recommendation is not to hold VMaccess for long periods of time
which is why the client must release VMaccess when doing a potential
long wait for server messages.

The code will be modified such that the client no longer acquires the
classUnloadMonitor and will no longer release VMaccess at the beginning of
a compilation. Instead, just as it is about to issue a read() operation
it will release VMaccess and re-acquire it immediately after.
During this time the GC could have unloaded some classes. Thus, after
getting VMaccess back we need to check the compilationShouldBeInterrupted
flag. If this flag is set we should abort the compilation with the proper
error code so that it can be retried.

This commmit introduces another small change: in shouldDowngradeCompReq
we will test whether a method is compiled with
entry->_oldStartPC == 0
instead of checking the startPC of the current method. This avoids a
message from the server to the client.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>