Initialize StringCompressionFlag when loading
When loading the StringCompressionFlag class initialize it as well as
the JIT depends on the initialization hook in jitHookClassPreinitialize
to capture the event and update its class hierarchy table.

When this table is updated all known assumptions made for this class
load will trigger and the corresponding side effect guards will be
patched at runtime. An example of this is the NOP side effect guards
placed for String compression.

Closes: #784
Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>