Do not send writeError to the server if it completed its task
During the client-server message exchange the client can discover
that the GC has unloaded classes and in that case it sends the server
a writeError message to abort current compilation.
However, this must not be done if the server just finished the compilation
because it may not listen anymore for a reply and this can deadlock
the client.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>