Fix interpreter performance regression in char[]-to-char[] decompressed array copy
As a performance optimization we directly call char[]-to-char[]
decompressedArrayCopy in a few places, as the method is JIT-recognized
and no bound checks are generated. However the interpreter implementation
of it is very slow, we need to implement native(s) similar to System.arraycopy
for all variants of String (de)compressedArrayCopy. As a stop-gap, for now
in Java 8, for char[]-to-char[] decompressedArrayCopy we call System.arraycopy
to address the -Xint DT startup regression.

Signed-off-by: Yan Luo <yan_luo@ca.ibm.com>