Fix BCDCHK call parameter evaluation order
Fix and refactor the BCDCHK evaluator so that, at the very
first stage of evaluation, the BCDCHK evaluator
evaluates the call parameter nodes that are not single reference
and not evaluated. The logic for skipping BCDCHK happens after this.

Doing this makes it easier to properly skip BCDCHK nodes that
have already been evaluated. Improper skipping of BCDCHK nodes
can cause RA to generate OOL sequence without a corresponding mainline
sequence. This can result in runtime memory corruptions or crashes.

Signed-off-by: Nigel Yu <yunigel@ca.ibm.com>