Fix race in JProfiling addKey
When rearranging values in a JProfiling hash table,
default matches with the first entry in the table
where possible at any point. This could result
in the table exceeding its capacity in the middle
of rearrangement. By ensuring the first slot in
the table holds -1 when not filled a default match
is not possible.

This also improves some of the JProfiling tracing
to debug mask calculation and rearrangement.

Signed-off-by: Nicholas Coughlin <cnic@ca.ibm.com>