Update the buffer length before calling decompress
The private move API only moves characters according to the capacity of
the current buffer. It does not update the length of the buffer itself.
Thus if we call the move API to reserve space for a character we are
about to insert and we are forced to decompress the buffer we must
update the length of the buffer prior to calling the decompress API so
that all characters in the compressed buffer are properly copied to the
decompressed buffer.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>