Allow jlink module info extension to be undone
In the Extensions for OpenJDK9 for OpenJ9, there is a change 
to the jlink module-info.java file which excludes the use of 
a Plugin that OpenJ9 doesn't support. 

Since undoing the module-info.java extension will require the 
existence of this class, I've implemented the only method in 
the Plugin abstract class that we need to implement, in a way 
that doesn't change anything (so the plugin is used, but should 
do nothing).

Signed-off-by: Adam Farley <adam.farley@uk.ibm.com>