Workaround: disable methodEnter/Exit hooks in JAAS mode
Communication between server and client is a major source of aveahed in
JAAS mode. Top 3 messages exchanged between server and client are
    VM_getClassFromSignature
    VM_isMethodEnterTracingEnabled
    VM_isMethodExitTracingEnabled

This commit eliminates the methodEnter/Exit queries by assuming that
methodEnter/Exit tracing is always disabled. This is an workaound
until we settle on some form of caching.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>