Co-operative copy of long object in CS
For Concurrent Scavenger, use copyOrWait() API to ensure that object is
fully copied before exposing the new verison of the object. The API is
no-op in Standard Scavenger.

Introduce a varion of getArrayLayout() that will return layout info
based on preserved info in ForwardedHeader, instead of looking into the
(destination) object header itself, which is not available with
co-operative copying, since the destination object header contains copy
progress info, while copy is still in progress.

Signed-off-by: Aleksandar Micic <amicic@ca.ibm.com>