Coerce the server to compile in parallel
The server is not very eager to take requests from the compilation queue due
to our normal heuristics: the activation of an additional compilation thread
happens only when the compilation queue grows over some limit.
Since we only have a client, there could only be 7 outstanding compilation
requests (one per client compilation thread) and this is not enough to
activate another thread at the server.
This commit adjusts the heuristics in JAAS_SERVER mode such that a request
is served as soon as it is added to the compilation queue.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>