Fix timing hole releasing safe point VM access
When releasing safe point VM access, the vmThreadListMutex must be
released after resetting the VM safePointState in order to prevent
newly-created threads from incorrectly setting the halted at safe point
bit.

The vmThreadListMutex is held for the entirety of the safe point, and it
is also required while creating a new thread, so new threads will not be
created during the safe point.  Releasing the vmThreadListMutex early
allows the thread creation code to proceed, where it examines the
safePointState and incorrectly set the halt bit which will never be
cleared, causing VM access acquisition to block forever.

Signed-off-by: Graham Chapman <graham_chapman@ca.ibm.com>