Introduce outgoing utf8Length parameter to copyStringToUTF8WithMemAlloc
Some of the callers call copyStringToUTF8WithMemAlloc to copy a string
and then subsequently call getStringUTF8Length to extract the length.
This is double computation as copyStringToUTF8WithMemAlloc already
computes this value. To avoid the double computation we add an outgoing
optional argument (the user can pass NULL if they don't care) to return
the length in number of bytes of the extracted UTF8 string excluding
the null terminator character.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>