Replace uses of copyStringToUTF8Helper with copyStringToUTF8WithMemAlloc
Most of the uses of copyStringToUTF8Helper are used in conjuction with
memory allocation and getStringUTF8Length. This is duplicated code
which copyStringToUTF8WithMemAlloc already handles. This also gives us
the ability to perform stack allocated buffer optimizations in the
places where we do not need dynamic allocation for the string we are
copying. This will be done in a subsequent commit.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>