Add prependStrLength parameter to copyStringToUTF8WithMemAlloc
Most of the time we use a compile time constant string as a prepend,
and usually an empty string. The compiler will usually not be able to
inline the copyStringToUTF8WithMemAlloc function and as such we will
incurr the call overhead of calling strlen. We push this computation
onus onto the caller and force the caller to pass the length of the
string to prepend (usually it will be zero).

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>