Remove utf8Len parameter from copyStringToUTF8Helper
For performance we delegate to the caller to ensure that the size of
the UTF8 buffer is large enough to contain the translated Unicode
string. With this optimization we can get rid of all checks inside of
the fast path for loop so the compiler can generate the fastest
possible sequence to encode the characters.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>