Allow opt level downgrades during grace period
This commit changes the existing heuristics for when the optimization level
of compilations is changed from 'warm' to 'cold'. Previously, such downgrades
were not allowed during the "grace period" (first second) to preserve behavior
of short running benchmarks. Recent experiments showed that we could further
improve startup of applications if we lift that restriction.
The new heuristic can be disabled with a new option:
  -Xjit:dontDowngradeToColdDuringGracePeriod

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>