Eliminate remote calls to isClassLibraryMethod
isClassLibraryMethod(TR_OpaqueMethodBlock *method, bool vettedForAOT = false)
is one of the most frequent remote calls from the server to client.
All calls to isClassLibraryMethod() originate from isMethodInValidLibrary()
which in turn is used during the constructor of a ResolvedJ9Method.
The solution adopted with this commit is add a new field
TR_ResolvedJ9JAASServerMethod::_isMethodInValidLibrary which will cache the
value sent by the client at the time it constructs the mirror ResolvedJ9Method.

After the implementation of this change I no longer see remote messages related
to isClassLibraryMethod

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>