Eliminate remote calls to setRecognizedMethodInfo/shouldFailSetRecognizedMethodInfoBecauseOfHCR
These calls are used during the construction of a TR_ResolvedJ9JAASServerMethod
object, more specifivally, they are used when calling construct().
All construct() does is to set the mandatoryRecognizedMethod and the
recognizedMethod fileds in TR_Method. However, these fields are already
computed by the mirror TR_ResolvedJ9Method created on the client, so
the client only has to send the server this information.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>