Remove spurious directories from mingw include path
Remove the INCLUDE environment variable from the mingw include path when
compiling OpenJ9.  The build system may set this variable to the MSVC include
path.  The result is double definitions of certain types and hence compilation
failures.  The mingw include path is sufficient without the INCLUDE variable.

In legacy builds retain this setting for compatibility.

Signed-off-by: Peter Bain <peter_bain@ca.ibm.com>