Allow configuring a JAAS timeout on the server
Similar to the client, the server can now cancel a compilation if
communication with the client is taking longer than a specified
threshold. This timeout is set via "-XX:jaasServer:timeout=<x>"
where <x> is the timeout duration in milliseconds.

The server stream needs to drain its completion queue after triggering
a cancellation due to a timeout event.

Signed-off-by: Omer Sheikh <Omer.Sheikh@ibm.com>