Implement client session support
On the server we want to persist data that is specific to a client so that
we don't have to ask the same questions when a compilation starts and to
segregate data from different clients when the server serves several clients
at the same time

Data specific to a client will be stored in a data structure of the form

class ClientSessionData
{
   int64_t _timeOfLastAccess;
   J9Class * _javaLangClassPtr;
 ...
   int8_t _inUse; // Number of concurrent compilations from the same client
};

We will keep a hashtable of clientID --> ClientSessionData mappings inside the
compInfo. This hashtable is global to the entire server and will be created
using persistent memory.

Access to the hashtable will be protected by the compilationMonitor.
When a compilation thread in the server starts processing a request, it gets
the clientID from the stream and, having acquired the compilation monitor,
it finds the corresponding ClientSessionData. A pointer to this entry is
stored in compInfoPerThread for easy access without a lock.
At the same time the _inUse flag is incremented on the entry and the
_timeOfLastAccess is updated with current time.
This pointer to the ClientSessionData is going to be reset when the
compilation ends and at the same time the_inUse flag is going to be
decremented.

The hashtable needs to purged periodically of stale entries.
A compilation thread checks the time since the last purge and if this exceeds
a certain value then, with compilation monitor in hand, it scans the entire
hashtable and finds entries that have not been accessed for a long time.
Such entries can be deleted from the hashtable.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>