Release VM access before sending the compilation request to server
I have encountered a deadlock scenario where the the client blocks in sending
the compilation requests. Because the client holds VM access this blocks GC
which in turn blocks other compilation threads that try to respond to
outstanding server queries which in turn PROBABLY blocks the server from
accepting the new compilation request.

The solution is to release VM access before sending the compilation
request and acquire it immediately after. We also need to check for
unloaded classes in this situation.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>