Reduce number of remote queries to getClassClassPointer
The solution involves using the client session data. One of the fields in the
client data is a J9Class pointer to java/lang/Class which is returned by
getClassClassPointer(). The first getClassClassPointer() query will retrieve
the information from the client and then cache it. Subsequent queries will
get the cached data avoiding the network roundtrip.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>