Refactor access check for processing methods
When looking up a method, various access checks occur. The existing
access checks on protected virtual methods produce a visibility non
module access error. When producing an error message, one case assumes
that the visibility error is a non module access error. Due to the
current code path, this has been successful. This commit explicitly
passes the type of the visibility error from the check to the creation
of the error message.


Signed-off-by: Talia McCormick <Talia.McCormick@ibm.com>