Reduce number of remote queries to isSubjectToPhaseChange
Eliminate the part that scans all methods in conatining class to find a
particular method called specInstance$.
Reorganize the code so that we can return false in vast majority
of instances without talking to the client.
We only need to send a query if we need to find out if the client is in
JIT startup phase.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>