Automate build instructions using Ansible
Ansible is the most popular open source automation tool on GitHub today
with more than a quarter million downloads per month.

The ansible playbook produces a docker image based on Ubuntu 16.4. The docker
image contains the Eclipse OpenJ9 virtual machine. The owner of the virtual
machine is the 'openj9' user (configurable) who is a member of the 'openj9' group
(configurable). A container created over the image runs as the 'openj9'
non-privileged user.

The build process is broken down into the following phases:
 - create a docker container with the build environment
 - checkout the sources of ibmruntimes, omr and openj9
 - build openj9 from sources
 - archive the jdk
 - create a docker image and put the jdk inside
 - clean-up the build-environment (the phase might be skipped)

To start the build do:

$  ansible-playbook -i inventories/local/hosts openj9-docker-image.yml

Signed-off-by: Pavel Samolysov <samolisov@gmail.com>