Export access to DDR for b165+
The existing code which reflectively exports access to DDR was
written prior to the refactoring which moved java.lang.reflect.Module
to java.lang.Module.

This code doesn't have the advantage of being preprocessed and so
can't make distinctions at that time. Instead, this version attempts
to locate Module in the new package first and then falls back to
looking in the old package.

Issue: #378
Signed-off-by: Keith W. Campbell <keithc@ca.ibm.com>