Adds jvmti class redefinition tests for nestmates
With the introduction of JEP 181, classes belong to a 'nest' - a group
of classes that share access between private fields. In order to define
a class's nest, class files gain two nestmates attributes: 'NestHost'
and 'NestMembers'. When redefining a class, there are restrictions on
the type of redefinition allowed. In particular, classes can not change
their nestmates attributes. This commit adds jvmti tests to check error
handling of nest attributes during class redefinition.

Signed-off-by: Talia McCormick <Talia.McCormick@ibm.com>