Various changes related to caching at the server
- Do not get the compilation monitor and try to process unloadedClasses unless
the list not empty. At the same place use 'findClientSession' rather than
'findOrCreateClientSession` because the latter can throw when allocating memory

- When a compilation thread on the server finds/creates the pertinent
ClientSessionData it also uses cacheRemoteROMClass. It's better to first check
whether the ROMClass already exists in the cache. Also, it's better to release
the compilationMonitor when using cacheRemoteROMClass because this routine
acquires another lock internally.

- Introduce a clientUID field for ClientSessionData for debugability.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>