Remove count field in String and use coder for compression
This change removes the count field from the String class in Java 9
onward. The String class in Java 8 is left unmodified. The count
field represents the length of the string in the underlying array.
Instead of using a separate field to represent the length we size the
underlying array to be the same length as the string.

By doing this we no longer have the ability to share the underlying
array with StringBuffer / StringBuilder objects (unless they are of
the correct size). Performance measurements have indicated this is a
worth while trade off.

Moreover, the elimination of the count field with it eliminates our
ability to use the sign bit of the count field to represent String
compression. As we still need to do this moving forward, and given
that Java 9 introduces a coder field we simply adopt to start using
this coder field to represent String compression instead.

To get to a performant state however our StringBuffer / StringBuilder
implementations also need to change. By introducing the coder field
we are more easily able to adapt to using OpenJDK's version of these
classes. This change also adopts the use of OpenJDK's StringBuffer /
StringBuilder for performance in Java 9 onward.

There were several steps made to make this change work. The original
commits were squashed ahead of time as development went on to ease the
pain of rebasing upstream changes to the files modified. The changes
made are outlined as follows:

- Separate String, StringBuffer, and StringBuilder preprocessor paths

Separate class definitions into contiguous chunks rather than
interleaved within the code. This makes it much simpler to remove the
count field for Java 9 and keep the Java 8 implementation intact.

- Temporarily disable JIT toUpper/toLower acceleration

These algorithms need to be adjusted to extract the length from the
right place. In fact I think these APIs need to be redesigned so these
implementation details are processed in the Java code. The helpers
should already receive the array and length as arguments.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>