Reduce number of remote queries for getClassFromSignature
This commit uses a cache that maintains <classLoader,className>--><J9Class>
mappings. The classloader is not the system class loader so entries in here
could be invalidated by class unloading. Because of this, this cache is
rebuilt during each compilation. At the beginning of the compilation the
cache is empty and it is populated from server-->client queries.
At the end of the compilation the cache is cleared.
Also, each compilation thread will have its own cache.

An alternative solution is to build a global cache that persists across
compilations. This cache needs to be scanned when the list of unloaded
classes is sent by the client. One problem is that the client informs
the server about classes that have been unloaded, but the key in this
hashtable is <classLoader,className>. Thus the client would have to send
the server a list of classloaders that have been unloaded as well.
Also, the cache needs to be organized like a hashtable of hashtables:
first search for a classloader and then search by name.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>