Add Proxy class's package to the defining classloader
j.l.reflect.Proxy bypasses ClassLoader when defining proxy classes.
Add the Proxy class's package to the classloader used to define the class.
If the classloader is null, use the platform (bootstrap) classloader.

Update Test_Package to account for change in behaviour between Java 8 and Java 9.
Fix a testNG warning.

Signed-off-by: Peter Bain <peter_bain@ca.ibm.com>