Add JaaS support for iProfiler method samples
This is the first properly working component out of many; the others
currently cause issues in the generated code or are otherwise not working,
so they are not included here.

The change to X86PrivateLinkage.cpp is required because of some issue with
PIC slots. It is a workaround and should be resolved better in the future.

This commit also includes a fix to the ILGenerator to avoid calling
getClassLoader.

Signed-off-by: Noah Weninger <noah.weninger@ibm.com>