Updates final field access check
Final fields in java must be set by the class in which it is declared.
Final instance variables must be set in every constructor of the class
they are declared in (<init>) and final static variables by the static
initializer of the class they are declared in (<clinit>).

Private fields/methods of a class can currently only be accessed by the
class in which they are declared. Earlier access checks therefore
assumed that checking if a field ref has private access to a final field
was sufficient.

With the advent of nestmates (see JEP 181), private fields/methods of a
class can be accessed by its fellow 'nest members' classes. The existing
access check would therefore allow two classes within the same 'nest' to
see eachothers' final fields within their own constructors.

This commit changes the access check from:

	if ( (has private access) && (is <init> or <clinit>) )

to be

	if ( (is same class) & (is <init> or <clinit>) )

Signed-off-by: Talia McCormick <Talia.McCormick@ibm.com>