Add shared classes support for nested jar
1. When constructing URL using recursiveJarTrim() for a nested jar, do
not trim the entry within the external jar so that it is kept in the
URL. For example, /path/A.jar!/lib/B.jar won't be trimmed to /path/A.jar
anymore.
2. Add a new API ClasspathEntryItem::getLocation() to return the path
for the external jar file. For example, in the case of
/path/A.jar!/lib/B.jar, getLocation() returns /path/A.jar, which will be
used in time stamping and stale marking operations. The full path
/path/A.jar!/lib/B.jar, returned by ClasspathEntryItem::getPath() will
be used in ClasspathEntryItem comparison, and trace/verbose/statistic
messages.

Signed-off-by: hangshao <hangshao@ca.ibm.com>