Allow j9javaArrayOfObject_load_VM() to be unused
j9javaArrayOfObject_load_VM() is defined inside a widely used header file.  If
included but not used it will cause a compilation failure.  Note that this does
not occur with -O3 for some reason.

Add an attribute to allow the function to be defined but not used (except on
Windows).

Minor drive-by comment cleanup.

Fixes https://github.com/eclipse/openj9/issues/1405

Signed-off-by: Peter Bain <peter_bain@ca.ibm.com>