WIP: Add jProfilingRecompilation test in loop
When jProfiling in compilation is enabled, add a test after aynchcheck node
in loop to calculate the frequency and trip up recompilation of the method
if reached certain threshold.

Following are the commit messages which deals with the changes requested as
per PR#1601 and fix to some issues found in personal testing.
Changes as per suggestions in PR#1601

This commit contains following changes requested on PR#1601
1. Common out the logic that creates a parent node to calculate
   raw frequency of block from JProfilingBlock and newly added
   JProfilingRecompLoopTest optimization.
2. Instead of getting a frequency of the region a block belongs
   to and use switch case to get the nesting depth, implement a
   method that walks the region struture counting loops to get the
   nesting depth.
3. Use dumpOptDetails instead of performTransformation.
4. Documentation about newly added API and some other newly added functions.
5. Use feGetEnv interface to set the base recompilation threshold.

Changes requested as per PR#1601 Part 2

This commit contains following changes as per suggestions on PR#1601
1. Instead of using getEnclosingBlock function on TreeTop which walk
   treetops backward till BBStart to get the enclosing block, store
   this information manually by checking for BBStart and add this
   information as well with TreeTop and loop depth in the test location
   information.

Changes requested as per PR#1601

This commit contains an addtional check in the function in BlockFrequencyInfo
API which is used to get the block number of original block particular BCI belongs
to at the time of creation of BlockFrequencyInfo where if someone tries to use the
ProfilingInfo in other compilation and requested BCI was not inlined previously.

Bug fix found in personal testing

Optimization scans for the asynccheck node while walking the treetops
and add them to list of test locations. This commit also keeps track of
all locations in extended basic blocks and avoids duplicate addition of
the test, i.e. if there are multiple asynccheck node associated with the
same bytecode info in extended basic blocks, we only need to put test
after very first asynccheck node amongst the ones with same bytecode info.

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>