Updates java.lang.Class API for new nestmates spec draft
Earlier spec draft for java.lang.Class nestmates API methods
'getNestHost' and 'getNestMembers' threw no exceptions and LinkageError
respectively. A newer spec draft allows both methods to throw a
SecurityException. This commit adds the necessary SecurityException for
this to occur.

Relevant portions of the newer spec draft include:

    public Class<?> getNestHost()
    ...
    Throws:
        SecurityException - If the returned class is not the current
class, and if a security manager, s, is present and the caller's class
loader is not the same as or an ancestor of the class loader for the
returned class and invocation of s.checkPackageAccess() denies access to
the package of the returned class
    ...
    public Class<?>[] getNestMembers()
    ...
    Throws:
    ...
        SecurityException - If any returned class is not the current
class, and if a security manager, s, is present and the caller's class
loader is not the same as or an ancestor of the class loader for that
returned class and invocation of s.checkPackageAccess() denies access to
the package of that returned class

Signed-off-by: Talia McCormick <Talia.McCormick@ibm.com>