Check return value from query_auxv before using it
There's a crash on PPC due to assuming query_auxv never fails.
The return value needs to be checked for -1 or 0 before using it
to map to something.

We can't check the AT_HWCAP queries as they may actually return
0 or -1 as the field being looked up is a bit field. Assuming
they succeed is the best we can do given the API.

Allow the AT_BASE_PLATFORM call to fail and default to UNKNOWN
if the AT_PLATFORM has already succeeded.

fixes: #1475

Signed-off-by: Dan Heidinga <daniel_heidinga@ca.ibm.com>