Update field offset calculation for classes with flattenable fields
Update instance field offset calculation to account for flattenable
value type fields. To find the offset of an instance field following a
flattened value type field, the instance size of the value type must be
known. For this reason, the `J9Class` of the value type is loaded in
`resolvefield.cpp`'s `fieldOffsetsFindNext()`.

Signed-off-by: Eric Zhang <eric99.zhang@gmail.com>