Fix a PPC TLE issue, a missing NULLCHK causing a crash in a transaction
When replacing a monitor-enter with a transaction we did not check if
the monitor-enter is under a Null-Check (i.e. a synchronized block) and
therefore we did not build the transaction with a preceding Null-Check.
This could allow for a SEGV during the transaction if the variable
being locked is null.
On PPC a SEGV within a transaction will result in a core dump rather
then a transaction rollback. For this reason we must have an explicit
Null-Check before the transaction starts. On platforms were a SEGV in
a transaction would result in a rollback, it still might be better to
Null-Check before the transaction to avoid a series of SEGV's during
the transaction attempts. At the moment TLE is only enabled on Z, so
this change is being done in case TLE is re-enabled on PPC in the
future.

Signed-off-by: Kevin Langman <langman@ca.ibm.com>